requirements.txt
# Core ML/RL libraries
gymnasium>=0.28
stable-baselines3>=2.2.1
torch>=2.0.0
tensorboard>=2.13.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
ta>=0.10.0

# Financial data
yfinance>=0.2.0
ccxt>=4.0.0

# Hyperparameter optimization
optuna>=3.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
python-binance>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Monitoring and logging
psutil>=5.9.0
schedule>=1.2.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Additional dependencies for enhanced features
joblib>=1.3.0
scipy>=1.10.0

# New dependency
sb3-contrib>=2.0.0